Integrate MoE layers_to_transform test into test_tuners_utils
Move the regression test for issue #3016 (layers_to_transform matching
the expert index instead of the layer index in MoE models) into the
existing TestTargetedModuleNames class and remove the standalone file.
Assert against the full expected set of targeted module names rather
than individual membership checks.
